made the worker not work

+++ db/migrate/20130708171312_create_category_scores.rb
@@ -1,9 +1,9 @@
 class CreateCategoryScores < ActiveRecord::Migration
 	def change
 		create_table :category_scores do |t|
-			t.integer :me, :default => 0, :null => false
-			t.integer :roommate, :default => 0, :null => false
-			t.integer :importance, :default => 0, :null => false
+			t.float :me, :default => 0, :null => false
+			t.float :roommate, :default => 0, :null => false
+			t.float :importance, :default => 0, :null => false
 			t.references :user
 			t.references :category
 		end
